db: move replayIngestedFlushable into recovery.go
